refactor: relocated simulation-related files into
