+++ scenarios/simple/sim/agent.py
@@ -3,13 +3,13 @@ import math
 import os
 from modules.utils import config, generate_positions 
 from modules.base_agent import BaseAgent
-from scenarios.simple.task import task_colors
+from scenarios.simple.sim.task import task_colors
 
 # Load agent configuration (Scenario Specific)
 work_rate = config['agents']['work_rate']
 
 # Load behavior tree
-behavior_tree_xml = f"{os.path.dirname(os.path.abspath(__file__))}/{config['agents']['behavior_tree_xml']}"
+behavior_tree_xml = f"{os.path.dirname(os.path.dirname(os.path.abspath(__file__)))}/{config['agents']['behavior_tree_xml']}"
 
 class Agent(BaseAgent):
     def __init__(self, agent_id, position, tasks_info):
