+++ scenarios/simple/sim/env.py
@@ -1,7 +1,7 @@
 from modules.base_env import BaseEnv
 from modules.utils import ResultSaver
-from scenarios.simple.task import generate_tasks
-from scenarios.simple.agent import generate_agents
+from scenarios.simple.sim.task import generate_tasks
+from scenarios.simple.sim.agent import generate_agents
 import pygame
 
 class Env(BaseEnv):
@@ -28,29 +28,30 @@ class Env(BaseEnv):
         self.data_records = []
 
     def save_results(self):
-        # Save gif
-        if self.save_gif and self.rendering_mode == "Screen":        
-            self.recording = False
-            print("Recording stopped.")
-            self.result_saver.save_gif(self.frames)          
+        pass
+        # # Save gif
+        # if self.save_gif and self.rendering_mode == "Screen":        
+        #     self.recording = False
+        #     print("Recording stopped.")
+        #     self.result_saver.save_gif(self.frames)          
      
 
-        # Save time series data
-        if self.save_timewise_result_csv:        
-            csv_file_path = self.result_saver.save_to_csv("timewise", self.data_records, ['time', 'agents_total_distance_moved', 'agents_total_task_amount_done', 'remaining_tasks', 'tasks_total_amount_left'])          
-            self.result_saver.plot_timewise_result(csv_file_path)
+        # # Save time series data
+        # if self.save_timewise_result_csv:        
+        #     csv_file_path = self.result_saver.save_to_csv("timewise", self.data_records, ['time', 'agents_total_distance_moved', 'agents_total_task_amount_done', 'remaining_tasks', 'tasks_total_amount_left'])          
+        #     self.result_saver.plot_timewise_result(csv_file_path)
         
-        # Save agent-wise data            
-        if self.save_agentwise_result_csv:        
-            variables_to_save = ['agent_id', 'task_amount_done', 'distance_moved']
-            agentwise_results = self.result_saver.get_agentwise_results(self.agents, variables_to_save)                        
-            csv_file_path = self.result_saver.save_to_csv('agentwise', agentwise_results, variables_to_save)
+        # # Save agent-wise data            
+        # if self.save_agentwise_result_csv:        
+        #     variables_to_save = ['agent_id', 'task_amount_done', 'distance_moved']
+        #     agentwise_results = self.result_saver.get_agentwise_results(self.agents, variables_to_save)                        
+        #     csv_file_path = self.result_saver.save_to_csv('agentwise', agentwise_results, variables_to_save)
             
-            self.result_saver.plot_boxplot(csv_file_path, variables_to_save[1:])
+        #     self.result_saver.plot_boxplot(csv_file_path, variables_to_save[1:])
 
-        # Save yaml
-        if self.save_config_yaml:                
-            self.result_saver.save_config_yaml()           
+        # # Save yaml
+        # if self.save_config_yaml:                
+        #     self.result_saver.save_config_yaml()           
    
     def record_timewise_result(self):
         agents_total_distance_moved = sum(agent.distance_moved for agent in self.agents)
